requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.44.1
blessed==1.20.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpickle==3.1.1
cmake==3.31.1
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.7
decorator==5.1.1
deepspeed==0.16.2
Deprecated==1.2.15
dill==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fairscale==0.4.13
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.7.2.post1
flatbuffers==25.1.24
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gco==2.0.2
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-pasta==0.2.0
gpustat==1.1
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
higher==0.2.1
hjson==3.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib-metadata==6.3.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
libclang==18.1.1
lit==18.1.8
loralib==0.1.2
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.5.1
matrix-client==0.4.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
ml-dtypes==0.4.1
mlconfig==0.1.8
modelscope==1.20.1
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx @ file:///croot/networkx_1717597493534/work
ninja==1.11.1.3
nltk==3.6.5
numpy==1.22.1
omegaconf==2.3.0
openai==0.27.9
opencv-python==4.8.0.76
opt_einsum==3.4.0
optree==0.14.0
oyaml==1.0
packaging==24.1
pandas==1.4.0
peft==0.7.1
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.28.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
pytorch-ranger==0.1.1
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.13.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
spandrel==0.4.0
sympy @ file:///croot/sympy_1724938189289/work
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.9.7
tokenizers==0.21.0
toolbox==1.11.0
torch==2.0.0+cu117
torch-optimizer==0.3.0
torchaudio==2.4.1
torchinfo==1.8.0
torchsde==0.2.6
torchvision==0.19.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.47.0
triton==2.0.0
typer==0.15.1
tzdata==2024.2
urllib3==1.26.20
wandb==0.19.1
warmup-scheduler-pytorch==0.1.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.17.0
xxhash==3.5.0
yarl==1.13.1
